[improve] improve resource management logic
