+++ streampark-console/streampark-console-service/src/main/java/org/apache/streampark/console/core/service/resource/ResourceHandle.java
@@ -1,0 +1,11 @@
+package org.apache.streampark.console.core.service.resource;
+
+import org.apache.streampark.console.base.domain.RestResponse;
+import org.apache.streampark.console.core.entity.Resource;
+
+public interface ResourceHandle {
+
+  RestResponse checkResource(Resource resourceParam) throws Exception;
+
+  void handleResource(Resource resource) throws Exception;
+}
